sep_backend/pyproject.toml
[tool.poetry]
name = "sep-backend"
version = "0.1.0"
description = "Backend for the Simple Event Planner app"
authors = ["Víctor Z. <desarrollo@proyectamos.cl>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "4.2.10"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
django-environ = "^0.11.2"
django-allauth = "0.57.0"
dj-rest-auth = {extras = ["with-social"], version = "^5.0.2"}
django-cors-headers = "^4.3.1"
drf-recaptcha = "^3.0.0"
django-recaptcha = "^4.0.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
